Changed dice roll item UI, bug fixes, refactoring
Replaced old DiceRoll list item view with simple TextView to to avoid focus problems with EditTexts in the ListView. User now changes num dice, dice size, and modifier in a dialog that opens after single clicking the list item.
When saving a dice set with the same name as an existing set, the user is asked whether to overwrite the old set or not.
Replaced the EditText in the save dice set dialog with an AutoCompleteTextView to make modifying old set easier.
If the user inputs invalid num dice or dice size values, Toasts inform the user of the error.
Refactored DatabaseHelper class